Вынес алгоритмы в отдельные файлы в строке и фибоначчи
