+++ src/types/types.ts
@@ -1,0 +1,6 @@
+import { ElementStates } from "./element-states";
+
+export type StringElement = {
+  value: string;
+  state: ElementStates;
+}
